+++ public/run.js
@@ -1,0 +1,3 @@
+/* globals window */
+window.Hercules.setUpChart();
+window.Hercules.setUpForm();
