+++ public/serverApi.js
@@ -1,12 +1,13 @@
 /* global window, fetch, Headers */
-window.Hercules.sendFormData = function sendFormData(params) {
+window.Hercules.postJSON = (json, url) => {
     // fetch is a Promise-based function which can be used instead of making
     // XMLHTTPRequest objects with callbacks. It is defined in most recent
     // browsers. Our code will only be tested in Chrome.
-    fetch("/crawl", {
+    return fetch(url, {
         method: "POST",
         headers: new Headers({ "Content-Type": "application/json" }),
-        body: JSON.stringify(params)
+        credentials: "same-origin", // enable cookies
+        body: JSON.stringify(json)
     }).then((response) => {
         if (!response.ok) {
             throw new Error("Server request returned " + response.statusText);
@@ -16,10 +17,8 @@ window.Hercules.sendFormData = function sendFormData(params) {
         // fetch throws a network error if there was an internet issue.
         // We throw an error if there was an invalid server response like 500.
         // Additionally response.json() throws an error with invalid JSON.
-        // In all cases, pretend we got a null response, which is ignored by cytoscape.
+        // In all cases, pretend we got a null response.
         console.error(err);
         return null;
-    }).then((json) => {
-        window.Hercules.populateChart(json);
     });
 };
